+++ com.alexander-yu-shamin.search-helper/Editor/SearchHelperSettings.cs
@@ -12,9 +12,11 @@ namespace SearchHelper.Editor
         public const string FindDependenciesToolName = "Find Dependencies";
         public const string FindUsedByToolName = "Find Used By";
         public const string FindByGuidToolName = "Find by GUID";
+        public const string FindUnusedToolName = "Find Unused";
 
         public const string ContextMenuItemFindDependenciesName = ContextMenuBase + FindDependenciesToolName;
         public const string ContextMenuFindUsedByItemName = ContextMenuBase + FindUsedByToolName;
         public const string ContextMenuShowObjectGuidItemName = ContextMenuBase + FindByGuidToolName;
+        public const string ContextMenuFindUnusedItemName = ContextMenuBase + FindUnusedToolName;
     }
 }